initial working window.
Needed to re-arrange project properties to get platform-specific
features working, including GLFW for window management. More details
to come.

Initial working demo -- window is open and you can close it. This is
not formatted into the project the way I want, however, so now we can
build this window into the architecture of the engine.