migrated opengl and windows-specific calls to create a graphics context
you can now make a graphics context by providing the type (e.g. opengl) and
the implementation defined by your compiler (e.g. Windows). The creation of
the context is still in the application, but should migrate to the renderer
and it's api once we have one.

Some basic OpenGL calls are made in the context just to prove it works; those
will also move to the Renderer API once we have one.